4.23 Terminado

+++ parte4/blogList/controllers/blogs.js
@@ -1,23 +1,14 @@
-const notesRouter = require('express').Router()
+const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const getTokenFrom = req => {
-  const authorization = req.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
-
-
-notesRouter.get('/', async (req, res) => {
+blogsRouter.get('/', async (req, res) => {
   const allBlogs = await Blog.find({}).populate('user', { username: 1, name: 1})
   res.json(allBlogs)
 })
 
-notesRouter.get('/:id', async (req, res, next) => {
+blogsRouter.get('/:id', async (req, res, next) => {
   const {id} = req.params
   try {
     const blog = await Blog.findById(id)
@@ -32,14 +23,12 @@ notesRouter.get('/:id', async (req, res, next) => {
   }
 }) 
 
-notesRouter.post('/', async (req, res, next) => {
+blogsRouter.post('/', async (req, res, next) => {
   const {title, author, url, likes} = req.body
   try{
-    const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET)
+    const user = await req.user
+    if (!user) return res.status(401).json({ error: 'Token invalid' })
 
-    if (!decodedToken.id) return res.status(401).json({ error: 'Token invalid' })
-
-    const user = await User.findById(decodedToken.id)
     const newBlog = new Blog({
       title,
       author,
@@ -47,7 +36,6 @@ notesRouter.post('/', async (req, res, next) => {
       likes,
       user: user.id
     })
-
     const result = await newBlog.save()
     user.blogs = user.blogs.concat(newBlog._id)
     await user.save()
@@ -59,13 +47,28 @@ notesRouter.post('/', async (req, res, next) => {
 
 })
 
-notesRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+blogsRouter.delete('/:id', async (req, res, next) => {
+  try {
+  const blog = await Blog.findById(req.params.id)
+  const user = req.user
+
+  if (!user) return res.status(401).json({ error: 'Token invalid' })
+  if (blog.user.toString() === user.id) {
+    user.blogs = user.blogs.filter(b => b._id.toString() !== req.params.id)
+    await user.save()
+    await Blog.findByIdAndDelete(req.params.id)
+    return res.status(204).end()
+  }
+  return res.status(401).json({ error: 'User invalid'})
+} catch(exception) {
+  next(exception)
+}
 })
 
-notesRouter.put('/:id', async (req, res) => {
+blogsRouter.put('/:id', async (req, res, next) => {
+  try {
   const { body, params } = req
+
 
   const updatedBlog = await Blog.findByIdAndUpdate(
     params.id,
@@ -73,7 +76,9 @@ notesRouter.put('/:id', async (req, res) => {
     {new: true, runValidators: true, context: 'query'})
     
   res.json(updatedBlog)
-  console.log(body, params.id)
+  } catch(exception) {
+    next(exception)
+  }
 })
 
-module.exports = notesRouter
+module.exports = blogsRouter
